Combine detectors into single library (#237)
* Make AtPxCT build again

* Allow targets to have many ROOT dictionaries

* Combine detectors into single library

This should help control the many root-level folders.

Adjusted the root dictionary generation to allow for multiple
dictionaries to be added to any target (library). This can speed up
build times for larger libraries if the dictionaries are split up
logically. The dictionaries are each defined by their own linkdefs.
See `AtDetectors` for an example.

* Combine AtField and AtPassive into AtDetectors

AtField and AtPassive are not used by anything else
in the compiled code base. They are loaded by ROOT
to run simulations at the same time as the detectors.

---------

Co-authored-by: anthoak13 <anthonya@frib.msu.edu>