Fix bug in decaying generators (#235)
* Refactor AtVertexPropagator

Instead of counting the event number with three
different variables, we now use a single variable
that is a boolean to distinguish between the
beam-like half of the event and the reacion-like
half of the event.

* Refactor AtVertexPropagator

Make the modification of the boolean more explicit
so the behavior is more clear.

* Add AtReactionGenerator as base for all reactions.

This class covers if the generator should be run
or not depending on which type of event we are
runnning.

It should also handle if there is a child
generator that still needs to run and will only
mark the reaction event complete if all generators
have been run.

* Remove redundant sequential decays

* Allow reaction gen to handle event type

* Fix reaction never generating

Because we are setting the event type at the end
of the generator phase (adding particles to the
primary track), we loose the information about
if it is reaction/beam in the stepping phase where
`ProcessHits` is called. To account for this we
use the trackID to determine what the event type
is. Alternatively, we could check
AtVertexPropogator we would just have to invert it

To account for hehavior in `AtTPC2Body` generator
added flag to `AtReactionGenerator` to allow for
any reaction to always run (even if it is a beam-
like event).

* Add additional tests

* Fix remaining generators

By definition, if an AtReactionGenerator is running
then it can create new particles. We should not
check for if it is a reaction event. That is
handled by the base class only and should be
hidden from users implementing generators.

* Fix formatting

* Fix clang-tidy warnings

* Refactor test system

- Move all tests to the same directory as the source files so they
are easier to locate in the project structure.

---------

Co-authored-by: anthoak13 <anthonya@frib.msu.edu>